Interceptor updates
