+++ csharp/msbuild/ice.test.sln
@@ -693,6 +693,18 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "client", "client", "{238FC0
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "server", "server", "{FA77217F-2B27-4434-8135-DCF177FE34A6}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "server", "server", "{63CF38CE-A34E-484B-83FA-B6D6594C257F}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "serveramd", "serveramd", "{1F2DCDC0-33F3-4B37-9DD0-BDB79B4FED51}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "collocated", "collocated", "{B6DBDAF8-B800-4BB9-A501-FDBED3261431}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "collocated", "..\test\Ice\interceptor\msbuild\collocated\collocated.csproj", "{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "serveramd", "..\test\Ice\interceptor\msbuild\serveramd\serveramd.csproj", "{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "server", "..\test\Ice\interceptor\msbuild\server\server.csproj", "{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -2657,6 +2669,54 @@ Global
 		{5F948ED7-8F5B-4CEE-81A7-E18A529F23BB}.Release|x64.Build.0 = Release|Any CPU
 		{5F948ED7-8F5B-4CEE-81A7-E18A529F23BB}.Release|x86.ActiveCfg = Release|Any CPU
 		{5F948ED7-8F5B-4CEE-81A7-E18A529F23BB}.Release|x86.Build.0 = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|Win32.Build.0 = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|x64.Build.0 = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Debug|x86.Build.0 = Debug|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|Win32.ActiveCfg = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|Win32.Build.0 = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|x64.ActiveCfg = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|x64.Build.0 = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|x86.ActiveCfg = Release|Any CPU
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614}.Release|x86.Build.0 = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|Win32.Build.0 = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|x64.Build.0 = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Debug|x86.Build.0 = Debug|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|Win32.ActiveCfg = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|Win32.Build.0 = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|x64.ActiveCfg = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|x64.Build.0 = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|x86.ActiveCfg = Release|Any CPU
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26}.Release|x86.Build.0 = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|Win32.Build.0 = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|x64.Build.0 = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Debug|x86.Build.0 = Debug|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|Win32.ActiveCfg = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|Win32.Build.0 = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|x64.ActiveCfg = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|x64.Build.0 = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|x86.ActiveCfg = Release|Any CPU
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -3006,6 +3066,12 @@ Global
 		{CCD6BE6C-D7E6-4A00-B0D3-C093F1220669} = {222498F0-8F68-4953-85B1-12D5D2A48082}
 		{238FC0C3-D8AD-47AC-878E-F5DFCD9D4B85} = {CCD6BE6C-D7E6-4A00-B0D3-C093F1220669}
 		{FA77217F-2B27-4434-8135-DCF177FE34A6} = {CCD6BE6C-D7E6-4A00-B0D3-C093F1220669}
+		{63CF38CE-A34E-484B-83FA-B6D6594C257F} = {73E908CB-B78C-4854-9959-BF49CEAC09FC}
+		{1F2DCDC0-33F3-4B37-9DD0-BDB79B4FED51} = {73E908CB-B78C-4854-9959-BF49CEAC09FC}
+		{B6DBDAF8-B800-4BB9-A501-FDBED3261431} = {73E908CB-B78C-4854-9959-BF49CEAC09FC}
+		{36FDA9E3-24CC-4C0B-8B92-19E1AB191614} = {B6DBDAF8-B800-4BB9-A501-FDBED3261431}
+		{1E6DC8FC-D5E4-4466-A713-2150CCB0AD26} = {1F2DCDC0-33F3-4B37-9DD0-BDB79B4FED51}
+		{AD0AE7C6-C8E1-4232-8CA1-061CF48EFD0C} = {63CF38CE-A34E-484B-83FA-B6D6594C257F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {647B3637-5175-460B-86AA-95BBD22CDE63}
