+++ csharp/src/Ice/Communicator.cs
@@ -1057,9 +1057,17 @@ namespace ZeroC.Ice
         public IEndpointFactory? IceFindEndpointFactory(Transport transport) =>
             _transportToEndpointFactory.TryGetValue(transport, out IEndpointFactory? factory) ? factory : null;
 
+        /// <summary>Adds one or more dispatch interceptors to the list of dispatch interceptors, the dispatch
+        /// interceptors register with a communicator will run after the interceptors register with an adapter, and
+        /// in the order that they were added.</summary>
+        /// <param name="interceptors">The dispatch interceptors to register with the communicator.</param>
         public void InterceptDispatch(params DispatchInterceptor[] interceptors) =>
             _dispatchInterceptors.AddRange(interceptors);
 
+        /// <summary>Adds one or more invocation interceptors to the list of invocation interceptors, the invocation
+        /// interceptors registered with the communicator are executed in the order they are registered for each
+        /// invocation done with the communicator.</summary>
+        /// <param name="interceptors">The invocation interceptors to register with the communicator.</param>
         public void InterceptInvocation(params InvocationInterceptor[] interceptors) =>
             _invocationInterceptors.AddRange(interceptors);
