+++ csharp/src/Ice/ObjectAdapter.cs
@@ -45,6 +45,8 @@ namespace ZeroC.Ice
 
         internal int IncomingFrameSizeMax { get; }
 
+        internal List<DispatchInterceptor> Interceptors { get; } = new List<DispatchInterceptor>();
+
         private static readonly string[] _suffixes =
         {
             "ACM",
@@ -85,7 +87,6 @@ namespace ZeroC.Ice
         private readonly Dictionary<string, IObject> _defaultServantMap = new Dictionary<string, IObject>();
         private int _directCount;  // The number of direct proxies dispatching on this object adapter.
         private TaskCompletionSource<object?>? _directDispatchCompletionSource;
-        private readonly List<DispatchInterceptor> _dispatchInterceptors = new List<DispatchInterceptor>();
         private Task? _disposeTask;
         private readonly string _id; // adapter id
         private readonly Dictionary<IdentityPlusFacet, IObject> _identityServantMap =
@@ -362,7 +363,7 @@ namespace ZeroC.Ice
             AddWithUUID("", servant, proxyFactory);
 
         public void Intercept(params DispatchInterceptor[] interceptors) =>
-            _dispatchInterceptors.AddRange(interceptors);
+            Interceptors.AddRange(interceptors);
 
         /// <summary>Removes a servant previously added to the Active Servant Map (ASM) using Add.</summary>
         /// <param name="identity">The identity of the Ice object.</param>
