+++ csharp/test/Ice/interceptor/.gitignore
@@ -1,0 +1,3 @@
+*.html
+*.xml
+*.xaml
