+++ csharp/test/Ice/interceptor/Collocated.cs
@@ -1,0 +1,25 @@
+//
+// Copyright (c) ZeroC, Inc. All rights reserved.
+//
+
+using System.Threading.Tasks;
+using Test;
+
+namespace ZeroC.Ice.Test.Interceptor
+{
+    public class Collocated : TestHelper
+    {
+        public override async Task RunAsync(string[] args)
+        {
+            await using Communicator communicator = Initialize(ref args);
+            communicator.SetProperty("TestAdapter.Endpoints", GetTestEndpoint(0));
+            ObjectAdapter adapter = communicator.CreateObjectAdapter("TestAdapter");
+            adapter.Add("test", new MyObject());
+            DispatchInterceptors.Register(adapter);
+            await adapter.ActivateAsync();
+            AllTests.Run(this);
+        }
+
+        public static Task<int> Main(string[] args) => TestDriver.RunTestAsync<Collocated>(args);
+    }
+}
