+++ csharp/test/Ice/interceptor/Server.cs
@@ -1,0 +1,27 @@
+//
+// Copyright (c) ZeroC, Inc. All rights reserved.
+//
+
+using System.Threading.Tasks;
+using Test;
+using ZeroC.Ice.Test.Interceptor;
+
+namespace ZeroC.Ice.Test.Operations
+{
+    public class Server : TestHelper
+    {
+        public override async Task RunAsync(string[] args)
+        {
+            await using Communicator communicator = Initialize(ref args);
+            communicator.SetProperty("TestAdapter.Endpoints", GetTestEndpoint(0));
+            ObjectAdapter adapter = communicator.CreateObjectAdapter("TestAdapter");
+            adapter.Add("test", new MyObject());
+            DispatchInterceptors.Register(adapter);
+            await adapter.ActivateAsync();
+            ServerReady();
+            await communicator.WaitForShutdownAsync();
+        }
+
+        public static Task<int> Main(string[] args) => TestDriver.RunTestAsync<Server>(args);
+    }
+}
