+++ csharp/test/Ice/interceptor/ServerAMD.cs
@@ -1,0 +1,26 @@
+//
+// Copyright (c) ZeroC, Inc. All rights reserved.
+//
+
+using System.Threading.Tasks;
+using Test;
+
+namespace ZeroC.Ice.Test.Interceptor
+{
+    public class ServerAMD : TestHelper
+    {
+        public override async Task RunAsync(string[] args)
+        {
+            await using Communicator communicator = Initialize(ref args);
+            communicator.SetProperty("TestAdapter.Endpoints", GetTestEndpoint(0));
+            ObjectAdapter adapter = communicator.CreateObjectAdapter("TestAdapter");
+            adapter.Add("test", new MyObjectAsync());
+            DispatchInterceptors.Register(adapter);
+            await adapter.ActivateAsync();
+            ServerReady();
+            await communicator.WaitForShutdownAsync();
+        }
+
+        public static Task<int> Main(string[] args) => TestDriver.RunTestAsync<ServerAMD>(args);
+    }
+}
