+++ csharp/test/Ice/interceptor/Test.ice
@@ -39,18 +39,7 @@ interface MyObject
 
     void op1();
 
-    // AMD version of the above:
-
-    // Simple add
-    [amd] int amdAdd(int x, int y);
-
-    // Will throw RetryException until current.Context["retry"] is "no"
-    [amd] int amdAddWithRetry(int x, int y);
-
-    // Throws remote exception
-    [amd] int amdBadAdd(int x, int y);
-
-    // Throws ONE
-    [amd] int amdNotExistAdd(int x, int y);
+    void shutdown();
 }
+
 }
