+++ csharp/test/Ice/interceptor/msbuild/collocated/collocated.csproj
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <AssemblyName>client</AssemblyName>
+    <AssemblyName>collocated</AssemblyName>
     <OutputType>Exe</OutputType>
     <UseAppHost>true</UseAppHost>
     <TargetFramework>$(AppTargetFramework)</TargetFramework>
@@ -9,7 +9,10 @@
   <ItemGroup>
     <Compile Include="../../../../TestCommon/TestHelper.cs" />
     <Compile Include="../../AllTests.cs" />
-    <Compile Include="../../Client.cs" />
+    <Compile Include="../../Collocated.cs" />
+    <Compile Include="..\..\DispatchInterceptors.cs" Link="DispatchInterceptors.cs" />
+    <Compile Include="..\..\TestAMDI.cs" Link="TestAMDI.cs" />
+    <Compile Include="..\..\TestI.cs" Link="TestI.cs" />
     <Compile Include="generated/Test.cs">
       <SliceCompileSource>../../Test.ice</SliceCompileSource>
     </Compile>
