+++ csharp/test/Ice/interceptor/msbuild/server/server.csproj
@@ -1,16 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <AssemblyName>client</AssemblyName>
+    <AssemblyName>server</AssemblyName>
     <OutputType>Exe</OutputType>
     <UseAppHost>true</UseAppHost>
     <TargetFramework>$(AppTargetFramework)</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="../../../../TestCommon/TestHelper.cs" />
-    <Compile Include="../../AllTests.cs" />
-    <Compile Include="../../Client.cs" />
-    <Compile Include="generated/Test.cs">
+    <Compile Include="../../Server.cs" />
+    <Compile Include="..\..\DispatchInterceptors.cs" Link="DispatchInterceptors.cs" />
+    <Compile Include="..\..\TestI.cs" Link="TestI.cs" />
+    <Compile Include="generated\Test.cs">
       <SliceCompileSource>../../Test.ice</SliceCompileSource>
     </Compile>
     <SliceCompile Include="../../Test.ice" />
