--- scripts/tests/Ice/interceptor.py
@@ -1,5 +1,0 @@
-#
-# Copyright (c) ZeroC, Inc. All rights reserved.
-#
-
-TestSuite(__name__, [ClientTestCase(client=Client(props={"Ice.Warn.Dispatch" : 0}))], libDirs=["interceptortest"])
